requirements.txt
accelerate
bitsandbytes
diffusers
transformers
torch
peft
numpy
Pillow
tqdm
PyYAML
datasets # For downloading the example dataset